dong some cleanup on unused code
