+++ recipe/views.py
@@ -19,6 +19,7 @@ from reportlab.platypus import *
 from reportlab.pdfbase import pdfmetrics
 from reportlab.pdfbase.ttfonts import TTFont
 from reportlab.pdfbase.pdfmetrics import registerFontFamily
+from taggit.models import Tag, TaggedItem
 import json
 
 
@@ -312,3 +313,14 @@ def recipeMail(request, id):
 class CookList(DetailView):
     model = Recipe
     template_name = "recipe/recipe_cook.html"
+
+
+def recipeTags(request, tag):
+    """displays a list of recipes with a giving tag"""
+    recipe_tag = get_object_or_404(Tag, slug=tag)
+    recipes_tagged = TaggedItem.objects.filter(tag=recipe_tag)
+    recipe_list = []
+    for recipe in recipes_tagged:
+        recipe_list.append(recipe.content_object)
+
+    return render_to_response('tags/recipe_tags.html', {'recipe_list': recipe_list}, context_instance=RequestContext(request))
