+++ settings.py
@@ -95,23 +95,24 @@ INSTALLED_APPS = (
     'django.contrib.admin',
     'django.contrib.flatpages',
     'django.contrib.staticfiles',
-    'grappelli.dashboard',
-    'grappelli',
-    'taggit',
-    #'taggit_templatetags',
-    'registration',
-    'rosetta',
-    'imagekit',
-    'haystack',
+
     'pagination',
-    'django_extensions',
-    'tastypie',
     'recipe',
     'recipe_groups',
     'ingredient',
     'accounts',
     'news',
     'list',
+
+    'grappelli.dashboard',
+    'grappelli',
+    'taggit',
+    'registration',
+    'rosetta',
+    'imagekit',
+    'haystack',
+    'django_extensions',
+    'tastypie',
 )
 
 # Password validation
