+++ urls.py
@@ -26,13 +26,11 @@ urlpatterns = (
     #(r'^profiles/', include('profiles.urls')),
     url(r'^rosetta/', include('rosetta.urls')),
     #(r'^follow/', include('relationships.urls')),
-    url(r'^friends/', include('friends.urls')),
     url(r'^feed/', include('feed.urls')),
     url(r'^groups/', include('recipe_groups.urls')),
     url(r'^recipe/', include('recipe.urls')),
     url(r'^ingredient/', include('ingredient.urls')),
     url(r'^list/', include('list.urls')),
-    url(r'^tags/', include('tags.urls')),
     url(r'^search/', include('haystack.urls')),
     url(r'^news/', include('news.urls')),
     url(r'^robots.txt$', TemplateView.as_view(template_name='robots.txt', content_type='text/plain')),
